Switch NullInt32 to NullInt64

+++ pkg/models/model_performer.go
@@ -118,8 +118,8 @@ type PerformerUrl struct {
 	URL         string    `db:"url" json:"url"`
 	Type        string    `db:"type" json:"type"`
 	ImageID uuid.NullUUID `db:"id" json:"image_id"`
-	Height  sql.NullInt32 `db:"height" json:"height"`
-	Width   sql.NullInt32 `db:"width" json:"width"`
+	Height  sql.NullInt64 `db:"height" json:"height"`
+	Width   sql.NullInt64 `db:"width" json:"width"`
 }
 
 func (p *PerformerUrl) ToURL() URL {
