+++ pkg/models/model_studio.go
@@ -55,8 +55,8 @@ type StudioUrl struct {
 	URL      string        `db:"url" json:"url"`
 	Type     string        `db:"type" json:"type"`
 	ImageID  uuid.NullUUID `db:"id" json:"image_id"`
-	Height   sql.NullInt32 `db:"height" json:"height"`
-	Width    sql.NullInt32 `db:"width" json:"width"`
+	Height   sql.NullInt64 `db:"height" json:"height"`
+	Width    sql.NullInt64 `db:"width" json:"width"`
 }
 
 func (p *StudioUrl) ToURL() URL {
@@ -66,8 +66,8 @@ func (p *StudioUrl) ToURL() URL {
 	}
     if p.ImageID.Valid && p.Height.Valid && p.Width.Valid {
         imageID := p.ImageID.UUID.String()
-        height := int(p.Height.Int32)
-        width := int(p.Width.Int32)
+        height := int(p.Height.Int64)
+        width := int(p.Width.Int64)
         url.ImageID = &imageID
         url.Height = &height
         url.Width = &width
